requirements.txt
torch
torchvision
transformers
accelerate
peft
bitsandbytes
sentencepiece
open_clip_torch
einops
gradio
wandb